requirements.txt
python-dotenv
matplotlib
httpx